Location GLTF State Cleanup (#10129)
